move `run_full_sync()` to `tests/`
